requirements.txt
openai
python-dotenv
json
requests
